+++ game/gui/widgets/gui_menu.cpp
@@ -75,6 +75,7 @@ void Gui::menu(Config &c) {
             ImGui::Checkbox("use light", &c.use_light);
             ImGui::Checkbox("use grayscale", &c.use_grayscale);
             ImGui::Checkbox("player light", &c.player_loses_light);
+            ImGui::Checkbox("collision boxes", &c.render_collision_boxes);
             ImGui::EndMenu();
         }
         //---------------------------